refactor(api): repository와 domain 패키지 분리
