+++ mopl-api/src/main/java/io/mopl/api/conversation/repository/ConversationParticipantRepository.java
@@ -1,5 +1,7 @@
-package io.mopl.api.conversation.domain;
+package io.mopl.api.conversation.repository;
 
+import io.mopl.api.conversation.domain.ConversationParticipant;
+import io.mopl.api.conversation.domain.ConversationParticipantId;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
