+++ mopl-api/src/main/java/io/mopl/api/conversation/repository/ConversationRepository.java
@@ -1,5 +1,6 @@
-package io.mopl.api.conversation.domain;
+package io.mopl.api.conversation.repository;
 
+import io.mopl.api.conversation.domain.Conversation;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
