+++ mopl-api/src/main/java/io/mopl/api/conversation/service/ConversationService.java
@@ -7,14 +7,14 @@ import io.mopl.api.content.service.ContentThumbnailUploadService;
 import io.mopl.api.conversation.domain.Conversation;
 import io.mopl.api.conversation.domain.ConversationParticipant;
 import io.mopl.api.conversation.domain.ConversationParticipantId;
-import io.mopl.api.conversation.domain.ConversationParticipantRepository;
-import io.mopl.api.conversation.domain.ConversationRepository;
 import io.mopl.api.conversation.domain.DirectMessage;
-import io.mopl.api.conversation.domain.DirectMessageRepository;
 import io.mopl.api.conversation.dto.ConversationDto;
 import io.mopl.api.conversation.dto.ConversationPage;
 import io.mopl.api.conversation.dto.ConversationSearchRequest;
 import io.mopl.api.conversation.dto.DirectMessageDto;
+import io.mopl.api.conversation.repository.ConversationParticipantRepository;
+import io.mopl.api.conversation.repository.ConversationRepository;
+import io.mopl.api.conversation.repository.DirectMessageRepository;
 import io.mopl.api.user.domain.User;
 import io.mopl.api.user.domain.UserRepository;
 import io.mopl.api.user.dto.UserSummary;
